Miscellaneous cleanup.
* Mark functions used only in same file as static.
* Remove unused 'encoding' argument from msgtowstr and wstrtoms (all
  callers passed NULL).
* Silence remaining compiler warnings on 64-bit Windows about assigning
  size_t to int without cast.
* Remove superfluous FORMAT_INT4/FORMAT_UINT4 defines.
* Remove useless dollar_quote argument from findTag (it was always '$'), and
  add comment.
* Turn OID_NAME into a #define, to avoid a useless sprintf.
* Turn the mapFuncs array into structs, for clarity, and add comment.

+++ pgtypes.c
@@ -1524,7 +1524,7 @@ getAtttypmodEtc(const StatementClass *stmt, int col, int *adtsize_or_longestlen)
 								}
 							}
 						}
-						*adtsize_or_longestlen += (maxscale << 16);
+						*adtsize_or_longestlen += (int) (maxscale << 16);
 					}
 				}
 			}
